feat: add update users
